test(function): exclude JSON_OVERLAPS allocation check under race
